fix(time): Make time property accessors return the correct values
Previously we used the playerPlugin as the source of currentTime and
duration information. However, due to the nature of various media
plugins it became apparent that the raw media element was ill suited for
requesting these properties. This patch changes the flow so that the
duration getter, and the currentTime getter/setter pair now go to active
plugin to fetch the relevant values.

Closes #36